Do not ignore the test errors in travis
From `man bash`
The return status of a pipeline is the exit status of the last command, unless the pipefail option is enabled.  If pipefail is enabled, the
pipeline's return status is the value of the last (rightmost) command to exit with a non-zero status, or zero if all commands exit success‐
fully.

- Do not tail the check output as the test log is now written in
  test-suit.log files
- On failure display the whole content of every test-suite.log file found
  in the build directory and exit with 1